# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Lunch Meet",
    "description": "Description"
  },
  "host": "3.35.54.135",
  "basePath": "/",
...
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "token": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header",
      "description": "로그인 이후 토큰을 여기다가 넣으세요~! Bearer 빠트리지 말구요 !"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/login": {
      "get": {
        "tags": ["Login"],
        "description": "로그인 정보 확인",
        "parameters": [],
        "security": [{ "token": [] }],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Login"],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/signup": {
      "post": {
        "tags": ["signup"],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nickname": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}